src/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.9.1
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==65.13.0
djangorestframework==3.16.1
drf-yasg==1.21.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.12.0
msgpack==1.1.2
mysqlclient==2.2.7
openai==2.8.1
packaging==25.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
